refactor(DateInput,Description): migrate stories
