+++ src/components/DateInput/__stories__/Disabled.stories.tsx
@@ -1,0 +1,8 @@
+import { Template } from './Template'
+
+export const Disabled = Template.bind({})
+
+Disabled.args = {
+  label: 'Disabled',
+  disabled: true,
+}
