+++ src/components/DateInput/__stories__/Playground.stories.tsx
@@ -1,0 +1,3 @@
+import { Template } from './Template'
+
+export const Playground = Template.bind({})
