+++ src/components/DateInput/__stories__/Required.stories.tsx
@@ -1,0 +1,8 @@
+import { Template } from './Template'
+
+export const Required = Template.bind({})
+
+Required.args = {
+  label: 'Required',
+  required: true,
+}
