+++ src/components/DateInput/__stories__/Template.tsx
@@ -1,0 +1,7 @@
+import { Story } from '@storybook/react'
+import { ComponentProps } from 'react'
+import DateInput from '..'
+
+export const Template: Story<ComponentProps<typeof DateInput>> = props => (
+  <DateInput {...props} />
+)
