+++ src/components/Description/__stories__/Selectable.stories.tsx
@@ -1,0 +1,18 @@
+import { Story } from '@storybook/react'
+import { ComponentProps } from 'react'
+import Description from '..'
+
+export const Selectable: Story<ComponentProps<typeof Description>> = () => (
+  <Description userSelect>
+    <Description.Term>Name</Description.Term>
+    <Description.Desc>Big instance</Description.Desc>
+    <Description.Term>Size</Description.Term>
+    <Description.Desc>125 GB</Description.Desc>
+  </Description>
+)
+
+Selectable.parameters = {
+  docs: {
+    storyDescription: 'Make description selectable using `selectable` prop.',
+  },
+}
